+++ CHANGELOG.md
@@ -9,6 +9,17 @@ Version 0.2.1.0
 <https://github.com/mstksg/conduino/releases/tag/v0.2.1.0>
 
 *   `hoistPipe` exported from *Data.Conduit*
+*   A handful of pipe primitive combinators added to *Data.Conduit*, including:
+    *   `feedbackPipe`
+    *   `zipPipe`
+    *   `&|` / `fuseBoth`
+    *   `|.` / `fuseUpstream`
+    *   `fuseBothMaybe`
+*   Some pipe runners added to *Data.Conduit*, including:
+    *   `squeezePipe` / `squeezePipeEither`
+    *   `feedPipe` / `feedPipeEither`
+*   *Data.Conduit.Lift* module, for working with monad transformers
+*   `iterM` added to *Data.Conduit.Combinators*
 
 Version 0.2.0.0
 ---------------